+++ packages/perseus-core/src/utils/split-perseus-item.test.ts
@@ -400,6 +400,12 @@ describe("splitPerseusItem", () => {
                             [0, 1],
                             [0, 1],
                         ],
+                        showAxisArrows: {
+                            xMin: true,
+                            xMax: true,
+                            yMin: true,
+                            yMax: true,
+                        },
                         graph: {type: "none"},
                         correct: {type: "none"},
                         lockedFigures: [],
@@ -427,6 +433,12 @@ describe("splitPerseusItem", () => {
                             [0, 1],
                             [0, 1],
                         ],
+                        showAxisArrows: {
+                            xMin: true,
+                            xMax: true,
+                            yMin: true,
+                            yMax: true,
+                        },
                         graph: {type: "none"},
                         showTooltips: false,
                         backgroundImage: {