+++ packages/perseus/src/widgets/expression/__docs__/expression.stories.tsx
@@ -4,7 +4,6 @@ import {ServerItemRendererWithDebugUI} from "../../../../../../testing/server-it
 import expressionExport from "../expression";
 import {expressionItem2, expressionItem3} from "../expression.testdata";
 
-import type {KeypadConfiguration} from "@khanacademy/perseus-core";
 import type {Meta, StoryObj} from "@storybook/react-vite";
 
 const meta: Meta = {
@@ -28,11 +27,6 @@ type Story = StoryObj<typeof ServerItemRendererWithDebugUI>;
 /** This story shows how the expression widget looks when the keypad is
  * configured with _every_ option it supports.  */
 export const DesktopKitchenSink = (args: Story["args"]): React.ReactElement => {
-    const keypadConfiguration: KeypadConfiguration = {
-        keypadType: "EXPRESSION",
-        extraKeys: ["x", "y", "z"],
-    };
-
     return (
         <div style={{padding: "2rem"}}>
             <expressionExport.widget
@@ -49,8 +43,16 @@ export const DesktopKitchenSink = (args: Story["args"]): React.ReactElement => {
                 userInput=""
                 trackInteraction={() => {}}
                 widgetId="expression"
-                keypadConfiguration={keypadConfiguration}
+                extraKeys={["x", "y", "z"]}
                 reviewMode={false}
+                answerForms={[
+                    {
+                        considered: "correct",
+                        form: false,
+                        simplify: false,
+                        value: "8675309",
+                    },
+                ]}
             />
         </div>
     );