+++ packages/perseus/src/widgets/interactive-graphs/interactive-graph.tsx
@@ -147,9 +147,16 @@ type RenderProps = {
      *
      * NOTE(kevinb): perseus_data.go defines this as Array<Array<number>>
      */
-    // TODO(kevinb): Add a transform function to interactive-graph.jsx to
-    // rename `range` to `ranges` so that things are less confusing.
     range: GraphRange;
+    /**
+     * Whether to show the arrows on the axis.
+     */
+    showAxisArrows: {
+        xMin: boolean;
+        xMax: boolean;
+        yMin: boolean;
+        yMax: boolean;
+    };
     /**
      * The type of graph
      */
@@ -178,6 +185,7 @@ type DefaultProps = {
     labels: string[];
     labelLocation: Props["labelLocation"];
     range: Props["range"];
+    showAxisArrows: Props["showAxisArrows"];
     step: Props["step"];
     backgroundImage: Props["backgroundImage"];
     markings: Props["markings"];
@@ -259,6 +267,12 @@ class InteractiveGraph extends React.Component<Props, State> {
             [-10, 10],
             [-10, 10],
         ],
+        showAxisArrows: {
+            xMin: true,
+            xMax: true,
+            yMin: true,
+            yMax: true,
+        },
         step: [1, 1],
         backgroundImage: defaultBackgroundImage,
         markings: "graph",
@@ -869,7 +883,6 @@ export default {
     name: "interactive-graph",
     displayName: "Interactive graph",
     widget: InteractiveGraph,
-    staticTransform: _.identity,
     getStartUserInput,
     getCorrectUserInput,
     getUserInputFromSerializedState,